refactor(maps): move date + icon helpers out of maps.helpers.js
Relocate getTimeAndDateFromModifiedTimestamp to shared/helpers.js and the button-icon lookups to maps.constants.js, reshaping the latter as constant maps instead of getter functions. Update all consumers.
